+++ packages/design-system/tests/e2e/specs/stories/button.e2e.js
@@ -2,11 +2,10 @@ describe('Button', () => {
   it('Should navigate through main navigation', () => {
     cy.visit('/#/web');
 
-    cy.get('#app aside nav a')
+    cy.get('#app aside nav p-link-pure')
       .contains('Button')
       .click()
-      .parents('a')
-      .should('have.class', 'router-link-active');
+      .should('have.attr', 'active', 'true');
 
     cy.get('#app main h1')
       .should('be.visible')