+++ packages/ui-kit-js/src/components/content/text-list/text-list/text-list.tsx
@@ -1,7 +1,7 @@
 import {JSX, Component, Prop, h, Host} from '@stencil/core';
 import cx from 'classnames';
 import {prefix} from '../../../../utils';
-import {TextColor} from '../../../../types';
+import { Theme } from '../../../../types';
 
 @Component({
   tag: 'p-text-list',
@@ -11,8 +11,11 @@ export class TextList {
   /** The type of the text list. */
   @Prop() public listType?: 'unordered' | 'ordered' = 'unordered';
 
-  /** Basic text list color variations. */
-  @Prop() public color?: TextColor = 'porsche-black';
+  /** Basic text color variations depending on theme property. */
+  @Prop() public color?: 'default' | 'neutral-1' | 'neutral-2' | 'neutral-3' | 'inherit' = 'default';
+
+  /** Adapts the text color depending on the theme. Has no effect when "inherit" is set as color prop. */
+  @Prop() public theme?: Theme = 'light';
 
   public render(): JSX.Element {
 
@@ -25,7 +28,8 @@ export class TextList {
     const textListClasses = cx(
       prefix('text-list'),
       prefix(`text-list--color-${this.color}`),
-      this.listType === 'ordered' && prefix('text-list--ordered')
+      this.listType === 'ordered' && prefix('text-list--ordered'),
+      this.color !== 'inherit' && prefix(`text-list--theme-${this.theme}`)
     );
 
     return (