+++ packages/ui-kit-react/projects/ui-kit-wrapper/CHANGELOG.md
@@ -4,15 +4,37 @@
 
 ### Change
 
+* Headline component
+  * Update color prop
+  * Add Theme prop
+  
+* Text component
+  * Update color prop
+  * Add Theme prop
+
 * Button component
   * New custom element
+  * Update color prop
+  * Add Theme prop
   
 * Button-Pure component
   * New custom element
- 
-* Remove button-regular and button-icon component
+  * Add Theme prop
+  
+* Text List component
+  * Update color prop
+  * Add Theme prop
+
+* Icon component
+  * Update color prop
+  * Add Theme prop
 
 * Link component
+  * New custom element
+  * Update color prop
+  * Add Theme prop
+  
+* Link-Pure component
   * New custom element
 
 * Text component
@@ -23,8 +45,18 @@
 * Text Link component
   * Reduce amount of text variants (defined in global interface)
 
+* Pagination component
+  * Update color prop
+  * Add Theme prop
+  
+* Spacing
+  * Remove global static spacing classes
+  * Remove responsive spacing classes (`a` to `g`)
+
+* Remove button-regular and button-icon component
+
 * Global Setup
-  * `porsche-ui-kit.css` style should not be imported by the application anymore
+  * `porsche-ui-kit.css` style should not be included by the application anymore
 
 ## [1.0.0-alpha.7] (2019-11-05)